+++ CHANGELOG.md
@@ -20,7 +20,11 @@ high state of flux, you're at risk of it changing without notice.
 - **Breaking Change** (with respect to `2.10.0-rc`)
   - `OptionT`
     - rename `match` to `matchE` and add `match`
+  - `TheseT`
+    - rename `match` to `matchE` and add `match`
   - `TaskOption`
+    - rename `match` / `matchW` to `matchE` / `matchWE` and add `match` / `matchW`
+  - `TaskThese`
     - rename `match` / `matchW` to `matchE` / `matchWE` and add `match` / `matchW`
 
 # 2.10.0-rc.5